LifeContingencies.jl

LifeContingencies.ConstantInterestRate

LifeContingencies.Decrement

LifeContingencies.FunctionalInterestRate

LifeContingencies.InterestRate

LifeContingencies.InterestRate

LifeContingencies.InterestRate

LifeContingencies.InterestRate

LifeContingencies.VectorInterestRate

LifeContingencies.A

LifeContingencies.A

LifeContingencies.C

LifeContingencies.D

LifeContingencies.M

LifeContingencies.N

LifeContingencies.P

LifeContingencies.V

LifeContingencies.l

LifeContingencies.rate

LifeContingencies.v

LifeContingencies.v

LifeContingencies.ä

LifeContingencies.ä

MortalityTables.omega

MortalityTables.p

MortalityTables.q

MortalityTables.ω

LifeContingencies.InterestRate — Type

InterestRate()

InterestRate is an abstract type that is the parent of various concrete implementations of interest rate calculations.

LifeContingencies.InterestRate — Method

InterestRate(f)

Construct a FunctionalInterestRate. Assumes that f is a function that takes a given time period and returns the annual effective rate for that period.

LifeContingencies.InterestRate — Method

InterestRate(v::Vector{Float64})

Construct a VectorInterestRate.

LifeContingencies.InterestRate — Method

InterestRate(i::Real)

Construct a ConstantInterestRate.

LifeContingencies.A — Method

A(lc::LifeContingency, duration, time)

Term insurance for n years for someone starting in the $x$th duration. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.A — Method

A(lc::LifeContingency, duration)

Whole life insurance for someone starting in the $x$th duration. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.C — Method

C(lc::LifeContingency, duration)

The $C_x$ actuarial commutation function where the duration argument is x. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.D — Method

D(lc::LifeContingency, time)

The $D_x$ actuarial commutation function where the time argument is x. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.M — Method

M(lc::LifeContingency, duration)

The $M_x$ actuarial commutation function where the duration argument is x. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.N — Method

N(lc::LifeContingency, duration)

The $N_x$ actuarial commutation function where the duration argument is x. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.P — Function

P(lc::LifeContingency,start_time=0)

A whole life insurance with 1 unit payable at the end of the year of death, and payable by net annual premiums, starting from start_time timepoint (often 0).

LifeContingencies.V — Function

V(lc::LifeContingency,t,start_time=0)

The net premium reserve at the end of year t, starting from time start_time (often 0).

LifeContingencies.l — Method

l(lc::LifeContingency, time)

The $l_x$ actuarial commutation function where the time argument is x. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.rate — Method

    rate(i::InterestRate,time)

The interst during time time.

LifeContingencies.v — Method

v(i::InterestRate, from_period, to_period)
v(i::InterestRate, period)

The three argument method returns the discount factor applicable between period from_period and to_period given InterestRate i. The two argument method returns the discount factor from period zero to period given InterestRate i.

Examples

julia> i = InterestRate(0.05)
julia> v(i,1)
0.9523809523809523
julia> v.(i,1:5)
5-element Array{Float64,1}:
 0.9523809523809523
 0.9070294784580498
 0.863837598531476
 0.8227024747918819
 0.7835261664684589
 julia> v(i,1,3)
0.9070294784580498

LifeContingencies.v — Method

v(i::InterestRate, period)

The discount rate at period period.

LifeContingencies.ä — Method

ä(lc::LifeContingency, duration,time)

Life annuity due for someone starting in the xth duration for time years.

To enter the ä character, type a and then \ddot. See more on how to input unicode in Julia.

LifeContingencies.ä — Method

ä(lc::LifeContingency, duration)

Life annuity due for someone starting in the their $x$th duration. Issue age is based on the issue_age in the LifeContingency lc.

To enter the ä character, type a and then \ddot. See more on how to input unicode in Julia.

MortalityTables.omega — Method

omega(i::InterestRate)

The last period that the interest rate is defined for. Assumed to be infinite (Inf) for functional and constant interest rate types. Returns the lastindex of the vector if a vector type. Also callable using ω instead of omega.

MortalityTables.p — Function

p(lc::LifeContingency,duration,time=1)
p(lc::LifeContingency,duration)

Return the probablity of survival for the given LifeContingency.

MortalityTables.q — Function

q(lc::LifeContingency,duration,time=1)
q(lc::LifeContingency,duration,time=1)

Return the probablity of death for the given LifeContingency.

MortalityTables.ω — Method

ω(lc::LifeContingency)

Returns the last defined period for both the interest rate and mortality table. In the future, this may only look up the omega of the mortality table.

LifeContingencies.ConstantInterestRate — Type

ConstantInterestRate()

ConstantInterestRate is a struct with a given rate that will act as the same rate for all periods. Construct by calling InterestRate() with a rate as an argument.

Examples

# 5% interest for all years
InterestRate()

LifeContingencies.Decrement — Type

Decrement()

Decrement is an abstract type to be the parent of decrements like mortality, lapse, etc.

LifeContingencies.FunctionalInterestRate — Type

FunctionalInterestRate()

FunctionalInterestRate is a struct with a rate_function that is a function that takes a time and returns an annual interst rate for that time. Construct by calling InterestRate() with a function as an argument.

Examples

# simply return 5% always
InterestRate(time -> 0.05) 

# for every period, return a normally distrubted rate
InterestRate((x -> rand(Normal(0.05, 0.01))))

# an autocorrelated rate
InterestRate(
        time -> time <= 1 ? 0.05 : rand(Normal(last(i5.rate_vector), 0.01)),
    )

LifeContingencies.VectorInterestRate — Type

VectorInterestRate()

VectorInterestRate is a struct with a given vector where the element t is the rate at time t. Construct by calling InterestRate() with a vector as an argument. Note that if you provide a short vector, you may inhibit other methods (e.g. commutation/insurance calculations) becuase you haven't defined interest rates for longer-dated periods.

Examples

# 5% interest for years 1, 2, and 3
InterestRate([0.05, 0.05, 0.05])