LifeContingencies.jl

LifeContingencies.Contingency

LifeContingencies.DiscountFactor

LifeContingencies.Frasier

LifeContingencies.InterestRate

LifeContingencies.JointAssumption

LifeContingencies.JointLife

LifeContingencies.LastSurvivor

LifeContingencies.LifeContingency

LifeContingencies.SingleLife

LifeContingencies.A

LifeContingencies.APV

LifeContingencies.C

LifeContingencies.D

LifeContingencies.M

LifeContingencies.N

LifeContingencies.P

LifeContingencies.V

LifeContingencies.disc

LifeContingencies.l

LifeContingencies.rate

LifeContingencies.ä

MortalityTables.cumulative_decrement

MortalityTables.omega

MortalityTables.survivorship

LifeContingencies.DiscountFactor — Type

DiscountFactor(int::InterestRate,time_step)

An iterator version of an interest rate which will generate a discount vector that is as long as need. This is useful because in the a lot of actuarial calculations, you have a variable ending point (omega) in a lot of calculations, but you just want the interest to expand as necessary to fit your vector.

timestep is the fractional portion of the rate period. E.g. if you are using annual rates, then time_step is fraction of the year in each of the itereated time steps.

Can also be constructed by passing a timestep to an InterestRate. See example below.

Examples

julia> time_step = 1;
julia> disc_factor = InterestRate(0.05)(time_step)

julia> using IterTools
julia> Iterators.take(df,5) |> collect
5-element Array{Any,1}:
 1.0
 0.9523809523809523
 0.9070294784580498
 0.863837598531476
 0.8227024747918819

LifeContingencies.Frasier — Type

Frasier()

The assumption of independnt lives in a joint life calculation. Is a subtype of JointAssumption.

LifeContingencies.InterestRate — Type

InterestRate(vector)
InterestRate(constant_rate)

InterestRate is an abstract type that is the parent of various concrete implementations of interest rate calculations.

Note that if you provide a short vector, you may inhibit other methods (e.g. commutation/insurance calculations) becuase you haven't defined interest rates for longer-dated periods.

Examples

julia> i = InterestRate([0.05, 0.05, 0.05])
julia> i = InterestRate(0.05)

LifeContingencies.JointLife — Type

struct JointLife
    lives
    contingency
    joint_assumption
end

A `Life` object containing the necessary assumptions for contingent maths related to a joint life insurance. Use with a `LifeContingency` to do many actuaral present value calculations.

Keyword arguments:

lives is a tuple of two SingleLifes

contingency default is LastSurvivor(). It is the trigger for contingent benefits. See ?Contingency.

joint_assumption Default value is Frasier(). It is the assumed relationship between the mortality of the two lives. See ?JointAssumption.

Examples

using MortalityTables
tbls = MortalityTables.tables()
mort = tbls["2001 VBT Residual Standard Select and Ultimate - Male Nonsmoker, ANB"]

l1 = SingleLife(
    mort       = mort.select[30], 
    issue_age  = 30          
)
l2 = SingleLife(
    mort       = mort.select[30], 
    issue_age  = 30          
)

jl = JointLife(
    lives = (l1,l2),
    contingency = LastSurvivor(),
    joint_assumption = Frasier()
)

LifeContingencies.LastSurvivor — Type

LastSurvivor()

The contingency whereupon benefits are payable upon both lives passing. Is a subtype of Contingency

LifeContingencies.LifeContingency — Type

struct LifeContingency
    life::Life

LifeContingencies.SingleLife — Type

struct SingleLife
    mort
    issue_age::Int
    alive::Bool
    fractional_assump::MortalityTables.DeathDistribution
end

A Life object containing the necessary assumptions for contingent maths related to a single life. Use with a LifeContingency to do many actuaral present value calculations.

Keyword arguments:

mort pass a mortality vector, which is an array of applicable mortality rates indexed by attained age

issue_age is the assumed issue age for the SingleLife and is the basis of many contingency calculations.

alive Default value is true. Useful for joint insurances with different status on the lives insured.

fractional_assump. Default value is Uniform(). This is a DeathDistribution from the MortalityTables.jl package and is the assumption to use for non-integer ages/times.

Examples

using MortalityTables
tbls = MortalityTables.tables()
mort = tbls["2001 VBT Residual Standard Select and Ultimate - Male Nonsmoker, ANB"]

SingleLife(
    mort       = mort.select[30], 
    issue_age  = 30          
)

LifeContingencies.A — Function

A(lc::LifeContingency,from_time=0,to_time=nothing)

Life insurance for someone starting at from_time and lasting until to_time. If to_time is nothing (the default), will be insurance until the end of the mortality table or interest rates.

Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.C — Method

C(lc::LifeContingency, to_time)

$C_x$ is a retrospective actuarial commutation function which is the product of the discount factor and the difference in l ($l_x$).

LifeContingencies.D — Method

D(lc::LifeContingency, to_time)

$D_x$ is a retrospective actuarial commutation function which is the product of the survivorship and discount factor.

LifeContingencies.M — Method

M(lc::LifeContingency, from_time)

The $M_x$ actuarial commutation function where the from_time argument is x. Issue age is based on the issue_age in the LifeContingency lc.

LifeContingencies.N — Method

N(lc::LifeContingency, from_time)

$N_x$ is a prospective actuarial commutation function which is the sum of the D ($D_x$) values from the given time to the end of the mortality table.

LifeContingencies.P — Method

P(lc::LifeContingency)
P(lc::LifeContingency,to_time)

The net premium for a whole life insurance (without second argument) or a term life insurance through to_time.

The net premium is based on 1 unit of insurance with the death benfit payable at the end of the year and assuming annual net premiums.

LifeContingencies.V — Method

V(lc::LifeContingency,time)

The net premium reserve at the end of year time.

LifeContingencies.disc — Method

disc(i::InterestRate, to_time)

The discount rate at time to_time.

LifeContingencies.l — Method

l(lc::LifeContingency, to_time)

$l_x$ is a retrospective actuarial commutation function which is the survivorship up to a certain point in time. By default, will have a unitary basis (ie 1.0), but you can specify basis keyword argument to use something different (e.g. 1000 is common in the literature.)

LifeContingencies.rate — Method

    rate(i::InterestRate,time)

The interst during time time.

LifeContingencies.ä — Function

ä(lc::LifeContingency, from_time=0,to_time=nothing)

Life annuity due for the life contingency lc with the benefit period starting at from_time and ending at to_time. If to_time is nothing, will be benefit until the end of the mortality table or interest rates.

To enter the ä character, type a and then \ddot. See more on how to input unicode in Julia.

MortalityTables.omega — Method

omega(lc::LifeContingency)
omega(l::Life)
omega(i::InterestRate)

Lifes and LifeContingencys

Returns the last defined timeperiod for both the interest rate and mortality table. Note that this is different than calling omega on a MortalityTable, which will give you the last `attainedage`.

Example: if the LifeContingency has issue age 60, and the last defined attained age for the MortalityTable is 100, then omega of the MortalityTable will be 100 and omega of the LifeContingency will be 40.

InterestRates

The last period that the interest rate is defined for. Assumed to be infinite (Inf) for functional and constant interest rate types. Returns the lastindex of the vector if a vector type.

MortalityTables.survivorship — Method

survivorship(lc::LifeContingency,from_time,to_time)
survivorship(lc::LifeContingency,to_time)

Return the probablity of survival for the given LifeContingency.

LifeContingencies.Contingency — Type

Contingency()

An abstract type representing the different triggers for contingent benefits. Available options to use include:

LastSurvivor()

LifeContingencies.JointAssumption — Type

JointAssumption()

An abstract type representing the different assumed relationship between the survivorship of the lives on a JointLife. Available options to use include:

Frasier()

LifeContingencies.APV — Method

APV(lc::LifeContingency,to_time)

The actuarial present value which is the survivorship times the discount factor for the life contingency.

MortalityTables.cumulative_decrement — Method

cumulative_decrement(lc::LifeContingency,to_time)
cumulative_decrement(lc::LifeContingency,from_time,to_time)

Return the probablity of death for the given LifeContingency.